Change bic loading
Turns out the previous loader was VERY flawed. It made Godot think that the scene file was not a scene and would not run properly. After several hours of tinkering I realized that there is a much easier way to do this so here it is. I should have tested more